Making the stock sample transactional

+++ src/test/java/test/chetan/integrationtest/SampleTestIT.java
@@ -25,17 +25,21 @@ public class SampleTestIT {
 	private Logger logger = LoggerFactory.getLogger(SampleTestIT.class);
 
 	private String stockServicePath = "sample-web/service/stockservice";
-	
-	@DataProvider(name="stocksToTest")
-	private Object [] [] stocksToTest() {
-		return new Object [] [] {
-				{"AAPL","Apple Corp"},
-				{"IBM","IBM Corp"},
-				{"NFLX","Netflix"},
-				
+
+	@DataProvider(name = "stocksToTest")
+	private Object[][] stocksToTest() {
+		return new Object[][] {
+				{ "AAPL", "Apple Corp", "I love apple let it pass", true },
+				{
+						"IBM",
+						"IBM Corp",
+						"They are now sucking. I am going to make this so long it will fail",
+						false },
+				{ "NFLX", "Netflix", "8 bucks is a great price", true },
+
 		};
 	}
-	
+
 	private WebClient createWebClient() {
 		JacksonJaxbJsonProvider jsonProvider = new JacksonJaxbJsonProvider();
 		ArrayList<JacksonJaxbJsonProvider> providers = new ArrayList<JacksonJaxbJsonProvider>();
@@ -44,51 +48,56 @@ public class SampleTestIT {
 		return client;
 	}
 
-
-	@Test(enabled = true, dataProvider="stocksToTest")
-	public void createStock(String ticker, String companyName) {
+	@Test(enabled = true, dataProvider = "stocksToTest")
+	public void createStock(String ticker, String companyName, String comments,
+			boolean expectedPass) {
 		// create the stock
 		WebClient client = createWebClient();
 		client.path(stockServicePath + "/stock");
 		StockDTO currentTestCase = new StockDTO();
 		currentTestCase.setCompanyName(companyName);
 		currentTestCase.setTickerSymbol(ticker);
+		currentTestCase.setComments(comments);
 		Response r = client.post(currentTestCase);
-		Assert.assertTrue(r.getStatus() == 200);
-
+		if (expectedPass) {
+			Assert.assertTrue(r.getStatus() == 200);
+		}
 		// now delete the stock
-		
+
 	}
 
-	@Test(enabled = true, dependsOnMethods = "createStock", dataProvider="stocksToTest")
-	public void getStock(String ticker, String companyName) {
-		
-		
+	@Test(enabled = true, dependsOnMethods = "createStock", dataProvider = "stocksToTest")
+	public void getStock(String ticker, String companyName, String comments,
+			boolean expectedPass) {
+
 		WebClient client = createWebClient();
 		client.path(stockServicePath + "/stock/{ticker}", ticker);
 		// client.accept("application/json");
 		Response r = client.get();
-		StockDTO stockJustAdded = r.readEntity(StockDTO.class);
+		if (expectedPass) {
 
-		
-		
-		Assert.assertEquals(stockJustAdded.getCompanyName(),
-				companyName);
-		Assert.assertEquals(stockJustAdded.getTickerSymbol(),
-				ticker);
+			StockDTO stockJustAdded = r.readEntity(StockDTO.class);
+
+			Assert.assertEquals(stockJustAdded.getCompanyName(), companyName);
+			Assert.assertEquals(stockJustAdded.getTickerSymbol(), ticker);
+		} else {
+			Assert.assertNotEquals(r.getStatus(), 200);
+		}
 
 	}
-	
-	@Test(enabled=true, dependsOnMethods="getStock", dataProvider="stocksToTest") 
-	public void deleteStock(String ticker, String companyName) {
-		
+
+	@Test(enabled = true, dependsOnMethods = "getStock", dataProvider = "stocksToTest")
+	public void deleteStock(String ticker, String companyName, String comments,
+			boolean expectedPass) {
+
 		WebClient deleteClient = createWebClient();
-		deleteClient.path(stockServicePath + "/stock/{ticker}",
-			ticker);
-		Response responseDelete = deleteClient.delete();
-		Assert.assertTrue(responseDelete.getStatus() == 200);
+		deleteClient.path(stockServicePath + "/stock/{ticker}", ticker);
+		//only delete stocks that were created in the first place
+		if (expectedPass) {
+			Response responseDelete = deleteClient.delete();
+			Assert.assertTrue(responseDelete.getStatus() == 200);
+		}
 
 	}
-	
 
 }
